Rename 'test' to 'tests' for consistency with convention
